backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.1
Flask-Limiter>=4.0
marshmallow>=3.0,<4.0
flask-marshmallow==1.2.1
marshmallow-sqlalchemy==1.1.0
python-dotenv==1.0.1
Werkzeug==3.1.3
gunicorn>=23.0
gevent>=24.0
pytest>=9.0
